# docs/openapi/delivery-plus-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Delivery Plus Public API",
    "version": "1.0.0",
    "description": "Public Gateway contract for the Delivery Plus food delivery platform. This specification exposes the canonical public paths behind the API Gateway and intentionally excludes internal-only service routes."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/users/me": {
      "get": {
        "summary": "Get the current authenticated user profile",
        "tags": [
          "users"
        ],
        "operationId": "users_get_api_users_me",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update the current authenticated user profile",
        "tags": [
          "users"
        ],
        "operationId": "users_patch_api_users_me",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/me/orders": {
      "get": {
        "summary": "Get the current user order history",
        "tags": [
          "users"
        ],
        "operationId": "users_get_api_users_me_orders",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get a user profile by id",
        "tags": [
          "users"
        ],
        "operationId": "users_get_api_users_id",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/me/avatar/image-upload-url": {
      "post": {
        "summary": "Create a presigned avatar upload URL",
        "tags": [
          "users"
        ],
        "operationId": "users_post_api_users_me_avatar_image_upload_url",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedImageUpload"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUploadUrlRequest"
              }
            }
          }
        }
      }
    },
    "/api/users/me/avatar/confirm": {
      "post": {
        "summary": "Confirm and save an uploaded avatar",
        "tags": [
          "users"
        ],
        "operationId": "users_post_api_users_me_avatar_confirm",
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageConfirmRequest"
              }
            }
          }
        }
      }
    },
    "/api/restaurants": {
      "post": {
        "summary": "Create a restaurant",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_post_api_restaurants",
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "List restaurants",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_get_api_restaurants",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": []
      }
    },
    "/api/restaurants/{id}": {
      "get": {
        "summary": "Get a restaurant by id",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_get_api_restaurants_id",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      },
      "patch": {
        "summary": "Update restaurant details",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_patch_api_restaurants_id",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/restaurants/{id}/status": {
      "patch": {
        "summary": "Update restaurant status",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_patch_api_restaurants_id_status",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/restaurants/{id}/ownership/{userId}": {
      "get": {
        "summary": "Check restaurant ownership",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_get_api_restaurants_id_ownership_userId",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/api/restaurants/{id}/image-upload-url": {
      "post": {
        "summary": "Create a presigned restaurant image upload URL",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_post_api_restaurants_id_image_upload_url",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedImageUpload"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestaurantImageUploadUrlRequest"
              }
            }
          }
        }
      }
    },
    "/api/restaurants/{id}/image-confirm": {
      "post": {
        "summary": "Confirm and save a restaurant image",
        "tags": [
          "restaurants"
        ],
        "operationId": "restaurants_post_api_restaurants_id_image_confirm",
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestaurantImageConfirmRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/orders/restaurant/{restaurantId}": {
      "get": {
        "summary": "List orders for a restaurant",
        "tags": [
          "orders"
        ],
        "operationId": "orders_get_api_orders_restaurant_restaurantId",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Bad Request"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ImageUploadUrlRequest": {
        "type": "object",
        "required": [
          "contentType"
        ],
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        }
      },
      "RestaurantImageUploadUrlRequest": {
        "type": "object",
        "required": [
          "imageType",
          "contentType"
        ],
        "properties": {
          "imageType": {
            "type": "string",
            "enum": [
              "cover",
              "logo"
            ]
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        }
      },
      "ImageConfirmRequest": {
        "type": "object",
        "required": [
          "objectKey"
        ],
        "properties": {
          "objectKey": {
            "type": "string"
          }
        }
      },
      "RestaurantImageConfirmRequest": {
        "type": "object",
        "required": [
          "imageType",
          "objectKey"
        ],
        "properties": {
          "imageType": {
            "type": "string",
            "enum": [
              "cover",
              "logo"
            ]
          },
          "objectKey": {
            "type": "string"
          }
        }
      },
      "PresignedImageUpload": {
        "type": "object",
        "required": [
          "uploadUrl",
          "objectKey",
          "expiresIn",
          "method",
          "fields"
        ],
        "properties": {
          "uploadUrl": {
            "type": "string",
            "format": "uri",
            "description": "POST multipart form data to this object-storage URL."
          },
          "objectKey": {
            "type": "string",
            "description": "Temporary key to submit to the corresponding confirmation endpoint."
          },
          "expiresIn": {
            "type": "integer",
            "example": 300
          },
          "method": {
            "type": "string",
            "enum": [
              "POST"
            ],
            "description": "HTTP method for the upload request."
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Include every field in the multipart form before adding the image file."
          }
        }
      }
    }
  }
}